+++ cedar-policy-symcc/CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 Cedar Language Version: TBD
 
+### Added
+
+- `Display` impl for `Env`
+
 ## [0.3.0] - 2026-02-17
 Cedar Language Version: 4.4