+++ cedar-policy-symcc/CHANGELOG.md
@@ -12,6 +12,10 @@ Cedar Language Version: TBD
 
 - `Display` impl for `Env`
 
+### Fixed
+
+- Fixed bug in `Term::interpret` caused by missing `bvsrem` case
+
 ## [0.3.0] - 2026-02-17
 Cedar Language Version: 4.4
 
@@ -43,7 +47,6 @@ use the performance-optimized primitives introduced in 0.2.0. (#2102)
 
 - Bug where returned counterexamples could occasionally be invalid by containing
 cycles in the entity data for entities irrelevant to the given policies (#2089)
-- Fixed bug in `Term::interpret` caused by missing `bvsrem` case
 
 ### Removed