+++ cedar-policy-symcc/CHANGELOG.md
@@ -12,6 +12,10 @@ Cedar Language Version: TBD
 
 - `Display` impl for `Env`
 
+### Fixed
+
+- Fixed bug in `Term::interpret` caused by missing `bvsrem` case
+
 ## [0.3.0] - 2026-02-17
 Cedar Language Version: 4.4