+++ cedar-policy-symcc/CHANGELOG.md
@@ -10,11 +10,16 @@ Cedar Language Version: TBD
 
 ### Added
 
-- `Display` impl for `Env`
+- `Display` impl for `Env` (#2182)
+
+### Changed
+
+- Refactored `Solver` trait to include `check_sat_with_model()` instead of `get_model()`,
+to make the API easier to use with fewer implicit ordering constraints (#2190)
 
 ### Fixed
 
-- Fixed bug in `Term::interpret` caused by missing `bvsrem` case
+- Fixed bug in `Term::interpret` caused by missing `bvsrem` case (#2185)
 
 ## [0.3.0] - 2026-02-17
 Cedar Language Version: 4.4