+++ cedar-policy-symcc/CHANGELOG.md
@@ -22,7 +22,7 @@ required when models are being enabled (#2192)
 
 ### Fixed
 
-- Fixed bug in `Term::interpret` caused by missing `bvsrem` case (#2185)
+- Fixed bug in `Term::interpret` caused by missing `bvsrem` case (#2185, #2203)
 
 ## [0.3.0] - 2026-02-17
 Cedar Language Version: 4.4